backend/requirements.txt
Django==4.2.9
psycopg2==2.9.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-rest-framework-docs==0.1.7
drf-yasg==1.21.7
Pillow
django-cors-headers
requests


asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.0.0
constantly==23.10.4
cryptography==42.0.1
daphne==4.0.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.9.0
zope.interface==6.1

channels_redis
aioredis
